Update AS conversion (part)

+++ megamek/src/megamek/common/UnitRole.java
@@ -356,7 +356,7 @@ public enum UnitRole {
 					|| unit.hasSPA(BattleForceSPA.MT)
 					|| unit.hasSPA(BattleForceSPA.CT)
 					|| unit.hasSPA(BattleForceSPA.ST)
-					|| (unit.hasSPA(BattleForceSPA.CT) && unit.getSPA(BattleForceSPA.CT) >= 50);
+					|| (unit.hasSPA(BattleForceSPA.CT) && (int)unit.getSPA(BattleForceSPA.CT) >= 50);
 		default:
 			break;
 		}
